(refactor) use composables as intermediary for types
